[FE] REFACTOR: Mapinfo 코드 리팩토링 #1137
